Add more info to gameid, such as teams and points
Change function to get_game_info and make it return a dataframe
containing high level info about each game. Not that duplicates arise so
that it will be easier to filter for info on a specific team.
Plan is to create a dictionary of GameID:boxscore.
Next to do is to think about how to store player specific data.