init commit for probably unused tool
Created this when I thought the class wasn't updating properly. Was
going to use it as a debuggin tool. Turns out I had put in the wrong
date, which is why the stats were coming up empty.
Leaving this tool here in case I want to use it later for something.
idk